+++ src/components/props/ParolangeloProps.tsx
@@ -1,0 +1,5 @@
+import {Parolangelo} from '../../rdf/types/Parolangelo'
+
+export default interface ParolangeloProps {
+    parolangelo: Parolangelo
+}