+++ src/components/props/PersonProps.tsx
@@ -1,5 +1,0 @@
-import {Vocaboliere} from '../../rdf/types/Vocaboliere'
-
-export default interface PersonProps {
-    person: Vocaboliere
-}