+++ src/components/props/VocaboliereProps.tsx
@@ -1,0 +1,5 @@
+import {Vocaboliere} from '../../rdf/types/Vocaboliere'
+
+export default interface VocaboliereProps {
+    vocaboliere: Vocaboliere
+}