+++ src/components/routes/contenuti/dashboard/ParolangeloCountSpotlight.tsx
@@ -6,15 +6,15 @@ import Spotlight from '../../../common/story/Spotlight'
 import {RDFStore} from '../../../../rdf/RDFStore'
 import {vocang} from '../../../../rdf/prefixes'
 
-export default function WordCountSpotlight() {
-    const [wordCount, setWordCount] = useState(0)
-    const [soloWordCount, setSoloWordCount] = useState(0)
+export default function ParolangeloCountSpotlight() {
+    const [parolangeloCount, setParolangeloCount] = useState(0)
+    const [soloParolangeloCount, setSoloParolangeloCount] = useState(0)
 
     useEffect(() => {
-        Parolangelo.all().then(concepts => {
-            setWordCount(concepts.length)
-            setSoloWordCount(concepts.filter(c =>
-                c.creators(node =>
+        Parolangelo.all().then(parolangelo => {
+            setParolangeloCount(parolangelo.length)
+            setSoloParolangeloCount(parolangelo.filter(p =>
+                p.creators(node =>
                     (RDFStore.store.any(node, undefined, vocang.namespace('Vocaboliere')) !== null)
                 )().length === 1
             ).length)
@@ -31,19 +31,19 @@ export default function WordCountSpotlight() {
         imageUrl={'/images/math.jpg'}
         imageAlt={'Matematica'}
     >
-        <h1> {wordCount} </h1>
+        <h1> {parolangeloCount} </h1>
         <p>
             Sono le <Link to={`${PAROLANGELO_ROUTE}/parolangelo`}>
                 parolangelo
             </Link> all&apos;interno del <strong> Vocabolangelo</strong>.<br/>
-            Di queste, <strong> {wordCount - soloWordCount} </strong> sono nate dalla collaborazione, mentre
-            <strong> {soloWordCount}</strong> sono state inventate dai singoli.<br/>
+            Di queste, <strong> {parolangeloCount - soloParolangeloCount} </strong> sono nate dalla collaborazione, mentre
+            <strong> {soloParolangeloCount}</strong> sono state inventate dai singoli.<br/>
             Quando ci saranno grandi traguardi tutti i <Link to={`${PAROLANGELO_ROUTE}/vocaboliere`}>
                 vocabolieri
             </Link> festeggeranno in qualche modo, ad esempio con un <Link to={`${PAROLANGELO_ROUTE}/festabolangelo`}>
                 festabolangelo
             </Link>.<br/>
-            Al momento puntiamo a raggiungere le <strong>{highestMultipleOf500(wordCount)}</strong> parolangelo!
+            Al momento puntiamo a raggiungere le <strong>{highestMultipleOf500(parolangeloCount)}</strong> parolangelo!
         </p>
         <ul className="actions stacked">
             <li>