+++ src/components/routes/parolangelo/ParolangeloLayout.tsx
@@ -10,38 +10,39 @@ import {Vocaboliere} from '../../../rdf/types/Vocaboliere'
 import ConditionalComponent from '../../common/conditional/ConditionalComponent'
 import InnerWrapper from '../../common/story/InnerWrapper'
 import {DateUtility} from '../../../util/DateUtility'
+import ParolangeloProps from '../../props/ParolangeloProps'
 
 export function ParolangeloLayout() {
 
-    const [concept, setConcept] =
+    const [parolangelo, setParolangelo] =
         useState<Parolangelo | undefined>(undefined)
     const params = useParams()
 
     useEffect(() => {
         RDFStore.safeCall(store => {
-            return new Parolangelo(store.sym(vocang.uri + params.conceptId))
-        }).then(concept =>
-            setConcept(concept)
+            return new Parolangelo(store.sym(vocang.uri + params.id))
+        }).then(parolangelo =>
+            setParolangelo(parolangelo)
         )
-    }, [params.conceptId])
+    }, [params.id])
 
-    if(concept !== undefined) {
+    if(parolangelo !== undefined) {
         return <Wrapper optionalModifiers={['divided']}>
             <InnerWrapper style={1}>
                 <header>
-                    <h1>{concept.prefLabel}</h1>
-                    <p>{concept.pronunciation}</p>
+                    <h1>{parolangelo.prefLabel}</h1>
+                    <p>{parolangelo.pronunciation}</p>
                 </header>
                 <div className="index align-left">
-                    <Definitions concept={concept}/>
-                    <Examples concept={concept}/>
-                    <Images concept={concept}/>
-                    <Videos concept={concept}/>
-                    <Synonyms concept={concept}/>
-                    <Related concept={concept}/>
-                    <Note concept={concept}/>
-                    <Created concept={concept}/>
-                    <Creators concept={concept}/>
+                    <Definitions parolangelo={parolangelo}/>
+                    <Examples parolangelo={parolangelo}/>
+                    <Images parolangelo={parolangelo}/>
+                    <Videos parolangelo={parolangelo}/>
+                    <Synonyms parolangelo={parolangelo}/>
+                    <Related parolangelo={parolangelo}/>
+                    <Note parolangelo={parolangelo}/>
+                    <Created parolangelo={parolangelo}/>
+                    <Creators parolangelo={parolangelo}/>
                 </div>
             </InnerWrapper>
         </Wrapper>
@@ -50,38 +51,34 @@ export function ParolangeloLayout() {
     }
 }
 
-interface ConceptSubLayoutProps {
-    concept: Parolangelo
-}
-
-function Definitions(props: ConceptSubLayoutProps){
+function Definitions(props: ParolangeloProps){
     return <NamedSection title={'Definizione'}>
         <List
             isOrdered={true}
-            list={props.concept.definitions}
+            list={props.parolangelo.definitions}
             elementContent={def => <p>{def}</p>}
         />
     </NamedSection>
 }
 
-function Examples(props: ConceptSubLayoutProps){
-    return <ConditionalComponent condition={() => props.concept.examples?.length > 0}>
+function Examples(props: ParolangeloProps){
+    return <ConditionalComponent condition={() => props.parolangelo.examples?.length > 0}>
         <NamedSection title={'Esempi'}>
             <List
                 isOrdered={false}
-                list={props.concept.examples}
+                list={props.parolangelo.examples}
                 elementContent={ex => <p>{ex}</p>}
             />
         </NamedSection>
     </ConditionalComponent>
 }
-function Creators(props: ConceptSubLayoutProps){
+function Creators(props: ParolangeloProps){
     const creatorId = (creator: Vocaboliere) => creator.node.RelativeUri(vocang)
     return <NamedSection title={'Vocabolieri'}>
         <List
             isOrdered={false}
             list={
-                props.concept.creators((node) =>
+                props.parolangelo.creators((node) =>
                     (RDFStore.store.any(node, undefined, vocang.namespace('Vocaboliere')) !== null)
                 )().map((node) => new Vocaboliere(node.node))
             }
@@ -94,16 +91,16 @@ function Creators(props: ConceptSubLayoutProps){
     </NamedSection>
 }
 
-function Images(props: ConceptSubLayoutProps){
-    return <ConditionalComponent condition={() => props.concept.images?.length > 0}>
+function Images(props: ParolangeloProps){
+    return <ConditionalComponent condition={() => props.parolangelo.images?.length > 0}>
         <NamedSection title={'Immagini'}>
             <List
                 styleNone={true}
                 isOrdered={false}
-                list={props.concept.images}
+                list={props.parolangelo.images}
                 elementContent={image =>
                     <span className="image left">
-                        <img src={image} alt={props.concept.prefLabel}/>
+                        <img src={image} alt={props.parolangelo.prefLabel}/>
                     </span>
                 }
             />
@@ -111,13 +108,13 @@ function Images(props: ConceptSubLayoutProps){
     </ConditionalComponent>
 }
 
-function Videos(props: ConceptSubLayoutProps){
-    return <ConditionalComponent condition={() => props.concept.videos?.length > 0}>
+function Videos(props: ParolangeloProps){
+    return <ConditionalComponent condition={() => props.parolangelo.videos?.length > 0}>
         <NamedSection title={'Video'}>
             <List
                 styleNone={true}
                 isOrdered={false}
-                list={props.concept.videos}
+                list={props.parolangelo.videos}
                 elementContent={video =>
                     <span className="image left">
                         <video width="50%" height="auto" autoPlay muted loop>
@@ -131,31 +128,32 @@ function Videos(props: ConceptSubLayoutProps){
     </ConditionalComponent>
 }
 
-function Created(props: ConceptSubLayoutProps){
-    if(props.concept.created !== null) {
+function Created(props: ParolangeloProps){
+    if(props.parolangelo.created !== null) {
         return <NamedSection title={'Data di creazione'}>
-            <p>{DateUtility.toDateString(new Date(props.concept.created.toString()))}</p>
+            <p>{DateUtility.toDateString(new Date(props.parolangelo.created.toString()))}</p>
         </NamedSection>
     } else {
         return <></>
     }
 }
 
-interface OtherConceptProps {
+interface OtherParolangeloProps {
     title: string
     condition: () => boolean
     list: Parolangelo[]
 }
-function OtherConcept(props: OtherConceptProps){
-    const conceptId = (concept: Parolangelo) => concept.node.RelativeUri(vocang)
+
+function OtherParolangelo(props: OtherParolangeloProps){
+    const id = (parolangelo: Parolangelo) => parolangelo.node.RelativeUri(vocang)
     return <ConditionalComponent condition={props.condition}>
         <NamedSection title={props.title}>
             <List
                 isOrdered={false}
                 list={props.list}
-                elementContent={concept =>
-                    <Link to={`/parolangelo/${conceptId(concept)}`}>
-                        <p>{concept.prefLabel}</p>
+                elementContent={parolangelo =>
+                    <Link to={`/parolangelo/${id(parolangelo)}`}>
+                        <p>{parolangelo.prefLabel}</p>
                     </Link>
                 }
             />
@@ -163,28 +161,28 @@ function OtherConcept(props: OtherConceptProps){
     </ConditionalComponent>
 }
 
-function Synonyms(props: ConceptSubLayoutProps){
-    return <OtherConcept
+function Synonyms(props: ParolangeloProps){
+    return <OtherParolangelo
         title={'Parolangelo Sinonimi'}
-        condition={() => props.concept.synonyms().length > 0}
-        list={props.concept.synonyms()}
+        condition={() => props.parolangelo.synonyms().length > 0}
+        list={props.parolangelo.synonyms()}
     />
 }
 
-function Related(props: ConceptSubLayoutProps){
-    return <OtherConcept
+function Related(props: ParolangeloProps){
+    return <OtherParolangelo
         title={'Parolangelo Correllate'}
-        condition={() => props.concept.related()?.length > 0}
-        list={props.concept.related()}
+        condition={() => props.parolangelo.related()?.length > 0}
+        list={props.parolangelo.related()}
     />
 }
 
-function Note(props: ConceptSubLayoutProps){
-    return <ConditionalComponent condition={() => props.concept.notes?.length > 0}>
+function Note(props: ParolangeloProps){
+    return <ConditionalComponent condition={() => props.parolangelo.notes?.length > 0}>
         <NamedSection title={'Note'}>
             <List
                 isOrdered={false}
-                list={props.concept.notes}
+                list={props.parolangelo.notes}
                 elementContent={note => <p>{note}</p>}
             />
         </NamedSection>